requirements.txt
fastprogress==1.0.3
geomloss==0.2.6
imageio==2.26.0
matplotlib==3.7.1
numpy==1.24.1
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
torch==1.10.1
torchvision==0.14.1
tqdm==4.65.0